test: extract dashboard fixture helpers
